+++ core/src/main/java/com/alibaba/fastjson2/JSONWriterUTF16.java
@@ -1796,7 +1796,7 @@ class JSONWriterUTF16
                 continue;
             }
 
-            long v = item.longValue();
+            long v = item;
             boolean writeAsString = nonStringAsString
                     || (browserCompatible && !(v <= 9007199254740991L && v >= -9007199254740991L));
             if (writeAsString) {
